IotPlatform resource filter mask corrected
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@14715 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtmdbx/CMaskContainer.cpp
@@ -432,7 +432,7 @@ void CMaskContainer::AddMask(IMask* mask, bool isInversion)
 {
 	QList<QPair<IMask*,quint64>>& targetList = isInversion ? m_maskListInv : m_maskList;
 
-	targetList.append(QPair<IMask*,int>(mask, 0));
+	targetList.append(QPair<IMask*,quint64>(mask, 0));
 }
